# Improves long-significand performance