--- a/include/fast_float/ascii_number.h
+++ b/include/fast_float/ascii_number.h
@@ -60,6 +60,7 @@ fastfloat_really_inline
 parsed_number_string parse_number_string(const char *p, const char *pend, chars_format fmt) noexcept {
   parsed_number_string answer;
   answer.valid = false;
+  answer.too_many_digits = false;
   answer.negative = (*p == '-');
   if ((*p == '-') || (*p == '+')) {
     ++p;
@@ -81,6 +82,8 @@ parsed_number_string parse_number_string(const char *p, const char *pend, chars_
         uint64_t(*p - '0'); // might overflow, we will handle the overflow later
     ++p;
   }
+  const char *const end_of_integer_part = p;
+
   int64_t exponent = 0;
   if ((p != pend) && (*p == '.')) {
     ++p;
@@ -111,9 +114,9 @@ parsed_number_string parse_number_string(const char *p, const char *pend, chars_
   int32_t digit_count =
       int32_t(p - start_digits); // used later to guard against overflows
   if(exponent > 0) {digit_count--;}
+  int64_t exp_number = 0;            // explicit exponential part
   if ((fmt & chars_format::scientific) && (p != pend) && (('e' == *p) || ('E' == *p))) {
     const char * location_of_e = p;
-    int64_t exp_number = 0;            // exponential part
     ++p;
     bool neg_exp = false;
     if ((p != pend) && ('-' == *p)) {
@@ -137,7 +140,8 @@ parsed_number_string parse_number_string(const char *p, const char *pend, chars_
         }
         ++p;
       }
-      exponent += (neg_exp ? -exp_number : exp_number);
+      if(neg_exp) { exp_number = - exp_number; }
+      exponent += exp_number;
     }
   } else {
     // If it scientific and not fixed, we have to bail out.
@@ -164,12 +168,29 @@ parsed_number_string parse_number_string(const char *p, const char *pend, chars_
     // We over-decrement by one when there is a decimal separator
     digit_count -= int(start - start_digits);
     if (digit_count > 19) {
-      answer.mantissa = 0xFFFFFFFFFFFFFFFF; // important: we don't want the mantissa to be used in a fast path uninitialized.
       answer.too_many_digits = true;
-      return answer;
+      // Let us start again, this time, avoiding overflows.
+      i = 0;
+      p = start_digits;
+      const uint64_t minimal_nineteen_digit_integer{1000000000000000000};
+      while((i < minimal_nineteen_digit_integer) && (p != pend) && is_integer(*p)) {
+        i = i * 10 + uint64_t(*p - '0');
+        ++p;
+      }
+      if (i >= minimal_nineteen_digit_integer) { // We have a big integers
+        exponent = end_of_integer_part - p + exp_number;
+      } else { // We have a value with a fractional component.
+          p++; // skip the '.'
+          const char *first_after_period = p;
+          while((i < minimal_nineteen_digit_integer) && (p != pend) && is_integer(*p)) {
+            i = i * 10 + uint64_t(*p - '0');
+            ++p;
+          }
+          exponent = first_after_period - p + exp_number;
+      }
+      // We have now corrected both exponent and i, to a truncated value.
     }
   }
-  answer.too_many_digits = false;
   answer.exponent = exponent;
   answer.mantissa = i;
   return answer;