If you have binary floating-point numbers, you ought to serialize them using no more than 17 digits. Anymore and you are being wasteful.  Our code is optimized for such a scenario. However, if someone has data with lots and lots of digits per floating-point number, then our performance is somewhat less. This PR is better if you expect to be processing long significand.

Configuration: AMD Rome Processor (GCC 10)

|  model |  before | after (this PR)  | gain/loss |
|---|---|---|---|
| uniform  |   1132.06 MB/s |  1121.70 MB/s  | -1% |
| uniform/concise  |  951.12 MB/s |   952.04 MB/s  | 0% |
| canada  |  802.83 MB/s |   824.92 MB/s  | +3% |
| big_ints  | 370.35 MB/s  |  520.40 MB/s  | + 40% |

See also https://github.com/fastfloat/fast_float/pull/27 for a different trade-off.